Menu de mejoras, y más cositas, como el formato de numeros con puntos, me apetecía vale?

+++ Assets/Scripts/SellableItem.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 public class SellableItem : GrabableItem {
 
-    public float sellAmount = 13f;
+    public long sellAmount = 13L;
     public bool insideGeode;
     public bool canBeSelled = true;
